# [moe training] use smaller block sizes for per group scaling kernels to improve perf