Stacked PRs:
 * #2698
 * #2697
 * #2696
 * #2671
 * #2669
 * __->__#2668


--- --- ---

### [moe training] use smaller block sizes for per group scaling kernels to improve perf

I noticed in a trace of fp8 rowwise MoE training for llama4, the triton kernel that does per token group scaling along dim0 was unexpectedly slow for some reason.

I took a look with NCU and found "grid too small" warning, meaning that I am not parallelizing sufficiently to utilize all SMs. 


<img width="1381" height="78" alt="Screenshot 2025-08-02 at 2 41 28 PM" src="https://github.com/user-attachments/assets/a369dce1-2ad7-4b3e-82c4-ad1324e3715d" />

To fix this, I adjusted the kernel autotuner config to use much smaller block sizes. 

Kernel benchmarking shows a 5x - 9x (!) speedup now:

Old autotuner configs
```
input_shape      n_groups  high_precision_dtype      torch_time_us    triton_time_us
-------------  ----------  ----------------------  ---------------  ----------------
(256, 4096)             4  torch.bfloat16                  1378.24           457.888
(256, 4096)             8  torch.bfloat16                  2806.5            264.992
(256, 4096)            16  torch.bfloat16                  5485.47           171.072
(4096, 4096)            4  torch.bfloat16                  1479.42           546.656
(4096, 4096)            8  torch.bfloat16                  2787.07           539.952
(4096, 4096)           16  torch.bfloat16                  5436.42           728.352
(65536, 4096)           4  torch.bfloat16                  7814.5           7410.94
(65536, 4096)           8  torch.bfloat16                  8674.78          7995.66
(65536, 4096)          16  torch.bfloat16                 10420.9           9111.55
```


New: autotuner configs
```
input_shape      n_groups  high_precision_dtype      torch_time_us    triton_time_us
-------------  ----------  ----------------------  ---------------  ----------------
(256, 4096)             4  torch.bfloat16                  1804.35            52.48
(256, 4096)             8  torch.bfloat16                  2934.3             53.568
(256, 4096)            16  torch.bfloat16                  5674.56            51.712
(4096, 4096)            4  torch.bfloat16                  1494.94            80.912
(4096, 4096)            8  torch.bfloat16                  2849.54            86.496
(4096, 4096)           16  torch.bfloat16                  5531.39           108.448
(65536, 4096)           4  torch.bfloat16                  7794.98          1049.63
(65536, 4096)           8  torch.bfloat16                  8661.28          1180.59
(65536, 4096)          16  torch.bfloat16                 10419             1603.23
```