@danielvegamyhre 2025-08-05
@zou3519 @oulgen i was told by Simon you might have some insight into what appears to be a bug in the `wrap_triton` api, see https://github.com/pytorch/ao/pull/2668#issuecomment-3155821808 for context. Could you please take a look?